# read-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the read-service API for querying anomalies and telemetry.",
        "title": "Anomaly Detection Read API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/anomalies": {
            "get": {
                "description": "List detected anomalies from the Postgres read model, filterable by various dimensions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "anomalies"
                ],
                "summary": "List detected anomalies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "From Timestamp",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "To Timestamp",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Source System",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Metric Name",
                        "name": "metric",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Region",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Instance ID",
                        "name": "instance",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.AnomalyResponse"
                            }
                        }
                    }
                }
            }
        },
        "/v1/predictions": {
            "get": {
                "description": "List degradation forecasts generated by the ML Predict Service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "predictions"
                ],
                "summary": "List degradation forecasts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/v1/series/state/{key}": {
            "get": {
                "description": "Retrieve detailed telemetry state (parameters and latest values) for a specific metric series by its key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Retrieve telemetry state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Composite Key: source|metric|env|region",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SeriesStateResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.AnomalyResponse": {
            "description": "A detected anomaly with its metric context and statistical boundaries.",
            "type": "object",
            "properties": {
                "baseline": {
                    "type": "number",
                    "example": 120
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-03-08T15:00:00Z"
                },
                "detector": {
                    "type": "string",
                    "example": "mad_detector"
                },
                "env": {
                    "type": "string",
                    "example": "production"
                },
                "event_id": {
                    "type": "string",
                    "example": "evt-12345"
                },
                "id": {
                    "type": "integer",
                    "example": 1024
                },
                "instance_id": {
                    "type": "string",
                    "example": "i-0abcd1234efgh5678"
                },
                "mad": {
                    "type": "number",
                    "example": 15.5
                },
                "metric_name": {
                    "type": "string",
                    "example": "latency"
                },
                "producer": {
                    "type": "string",
                    "example": "pattern-service"
                },
                "region": {
                    "type": "string",
                    "example": "us-east-1"
                },
                "source": {
                    "type": "string",
                    "example": "payment-gateway"
                },
                "threshold": {
                    "type": "number",
                    "example": 300
                },
                "threshold_k": {
                    "type": "number",
                    "example": 3
                },
                "trace_id": {
                    "type": "string",
                    "example": "trace-550e8400-e29b-41d4-a716-446655440000"
                },
                "ts": {
                    "type": "integer",
                    "example": 1709400000
                },
                "value": {
                    "type": "number",
                    "example": 350.5
                },
                "window_size": {
                    "type": "integer",
                    "example": 60
                }
            }
        },
        "dto.SeriesStateResponse": {
            "description": "The latest statistical state for a specific metric series.",
            "type": "object",
            "properties": {
                "baseline": {
                    "type": "number",
                    "example": 120
                },
                "detector": {
                    "type": "string",
                    "example": "mad_detector"
                },
                "env": {
                    "type": "string",
                    "example": "production"
                },
                "last_ts": {
                    "type": "integer",
                    "example": 1709400000
                },
                "last_value": {
                    "type": "number",
                    "example": 125.5
                },
                "mad": {
                    "type": "number",
                    "example": 15.5
                },
                "metric_name": {
                    "type": "string",
                    "example": "latency"
                },
                "producer": {
                    "type": "string",
                    "example": "ingestion-service"
                },
                "region": {
                    "type": "string",
                    "example": "us-east-1"
                },
                "source": {
                    "type": "string",
                    "example": "payment-gateway"
                },
                "threshold": {
                    "type": "number",
                    "example": 300
                },
                "threshold_k": {
                    "type": "number",
                    "example": 3
                },
                "trace_id": {
                    "type": "string",
                    "example": "trace-550e8400-e29b-41d4-a716-446655440000"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-03-08T15:00:00Z"
                },
                "window_size": {
                    "type": "integer",
                    "example": 60
                }
            }
        }
    }
}